+++ android/sdk/src/main/java/com/tencent/mtt/hippy/HippyEngineManagerImpl.java
@@ -558,7 +558,7 @@ public abstract class HippyEngineManagerImpl extends HippyEngineManager implemen
 		}
 		resetEngine();
 
-		mEngineContext = new HippyEngineContextImpl(mDebugMode);
+		mEngineContext = new HippyEngineContextImpl(mDebugMode, mServerHost);
 		mEngineContext.getBridgeManager().initBridge(new Callback<Boolean>()
 		{
 			@Override
@@ -759,10 +759,11 @@ public abstract class HippyEngineManagerImpl extends HippyEngineManager implemen
 		 */
 		private DomManager													mDomManager;
 
-		public HippyEngineContextImpl(boolean isDevModule)
+		public HippyEngineContextImpl(boolean isDevModule, String debugServerHost)
 		{
 			mModuleManager = new HippyModuleManagerImpl(this, mAPIProviders);
-			mBridgeManager = new HippyBridgeManagerImpl(this, mCoreBundleLoader, HippyEngineManagerImpl.this.getBridgeType(), mEnableHippyBuffer, isDevModule, mGroupId, mThirdPartyAdapter);
+			mBridgeManager = new HippyBridgeManagerImpl(this, mCoreBundleLoader, HippyEngineManagerImpl.this.getBridgeType(),
+					mEnableHippyBuffer, isDevModule, debugServerHost, mGroupId, mThirdPartyAdapter);
 			mRenderManager = new RenderManager(this, mAPIProviders);
 			mDomManager = new DomManager(this);
 		}