+++ packages/hippy-vue/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.5](https://github.com/Tencent/Hippy/tree/master/packages/hippy-vue/compare/2.1.4...2.1.5) (2020-12-10)
+
+**Note:** Version bump only for package @hippy/vue
+
+
+
+
+
+## [2.1.4](https://github.com/Tencent/Hippy/tree/master/packages/hippy-vue/compare/2.1.3...2.1.4) (2020-12-03)
+
+
+### Features
+
+* **hippy-vue,hippy-react:** added setNativeProps on element ([#430](https://github.com/Tencent/Hippy/tree/master/packages/hippy-vue/issues/430)) ([d1f7e21](https://github.com/Tencent/Hippy/tree/master/packages/hippy-vue/commit/d1f7e216b5fef46ace0cf50803ad2940b429a0d6))
+* **hippy-vue,hippy-react:** perf setNativeProps ([5cd1291](https://github.com/Tencent/Hippy/tree/master/packages/hippy-vue/commit/5cd12910262ad3bb15d07c2dc974a829958a2b86))
+
+
+
+
+
 ## [2.1.2](https://github.com/Tencent/Hippy/tree/master/packages/hippy-vue/compare/2.1.1...2.1.2) (2020-11-23)